format and lint fixes
