--- apps/plex-recently-added/manifest.yaml
@@ -1,8 +1,0 @@
----
-id: plex-recently-added
-name: Plex Recently Added
-summary: Display Plex's recently added
-desc: Displays recently added on Plex server. Recommended to set up a local proxy server `index.js` to host the data. See <link here> for more information. See README.md for more information.
-author: noahpodgurski
-fileName: plex-recently-added.star
-packageName: plex-recently-added
