+++ apps/plexrecentlyadded/manifest.yaml
@@ -1,0 +1,8 @@
+---
+id: plex-recently-added
+name: PlexRecentlyAdded
+summary: Display Plex recently added
+desc: Displays recently added on Plex server. Recommended to set up a local proxy server 'index.js' to host the data. See <link here> for more information. See README.md for more information.
+author: noahpodgurski
+fileName: plex_recently_added.star
+packageName: plexrecentlyadded
